adding invoice using pdfkit

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 const { validationResult } = require('express-validator/check')
 
 
@@ -169,6 +170,7 @@ exports.getAddProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       if(image){
+      fileHelper.deleteFile(product.imageUrl);
       product.imageUrl = image.path;
       }
       product.description = updatedDesc;
@@ -186,17 +188,23 @@ exports.getAddProduct = (req, res, next) => {
 
   
 
-  exports.postDeleteProduct = (req, res, next) =>{
+  exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId : req.user._id })
-    .then( () => {
-      console.log("Destory product Sucessfully...");
-      res.redirect('/admin/products');
-    })
-    .catch( err => {
-      const error = new Error(err);
+    Product.findById(prodId)
+      .then(product => {
+        if (!product) {
+          return next(new Error('Product not found.'));
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({ _id: prodId, userId: req.user._id });
+      })
+      .then(() => {
+        console.log('DESTROYED PRODUCT');
+        res.redirect('/admin/products');
+      })
+      .catch(err => {
+        const error = new Error(err);
         error.httpStatusCode = 500;
         return next(error);
-    }); 
-    
-  }
+      });
+  };
